requirements.txt
Pillow
PySimpleGUI==4.60.5
